refactor(filter control): rename related components
